Xcode 5.0 (5A1413)

+++ UIKit.framework/iPhoneOS.platform/UICollectionViewDataSource-Protocol.h
@@ -10,10 +10,10 @@
 
 @protocol UICollectionViewDataSource <NSObject>
 - (UICollectionViewCell *)collectionView:(UICollectionView *)arg1 cellForItemAtIndexPath:(NSIndexPath *)arg2;
-- (int)collectionView:(UICollectionView *)arg1 numberOfItemsInSection:(int)arg2;
+- (long long)collectionView:(UICollectionView *)arg1 numberOfItemsInSection:(long long)arg2;
 
 @optional
 - (UICollectionReusableView *)collectionView:(UICollectionView *)arg1 viewForSupplementaryElementOfKind:(NSString *)arg2 atIndexPath:(NSIndexPath *)arg3;
-- (int)numberOfSectionsInCollectionView:(UICollectionView *)arg1;
+- (long long)numberOfSectionsInCollectionView:(UICollectionView *)arg1;
 @end
