+++ UIKit.framework/iPhoneOS.platform/UIPickerViewDataSource-Protocol.h
@@ -9,7 +9,7 @@
 @class UIPickerView;
 
 @protocol UIPickerViewDataSource <NSObject>
-- (int)pickerView:(UIPickerView *)arg1 numberOfRowsInComponent:(int)arg2;
-- (int)numberOfComponentsInPickerView:(UIPickerView *)arg1;
+- (long long)pickerView:(UIPickerView *)arg1 numberOfRowsInComponent:(long long)arg2;
+- (long long)numberOfComponentsInPickerView:(UIPickerView *)arg1;
 @end
