+++ UIKit.framework/iPhoneOS.platform/UIPickerViewDelegate-Protocol.h
@@ -11,11 +11,11 @@
 @protocol UIPickerViewDelegate <NSObject>
 
 @optional
-- (void)pickerView:(UIPickerView *)arg1 didSelectRow:(int)arg2 inComponent:(int)arg3;
-- (UIView *)pickerView:(UIPickerView *)arg1 viewForRow:(int)arg2 forComponent:(int)arg3 reusingView:(UIView *)arg4;
-- (NSAttributedString *)pickerView:(UIPickerView *)arg1 attributedTitleForRow:(int)arg2 forComponent:(int)arg3;
-- (NSString *)pickerView:(UIPickerView *)arg1 titleForRow:(int)arg2 forComponent:(int)arg3;
-- (float)pickerView:(UIPickerView *)arg1 rowHeightForComponent:(int)arg2;
-- (float)pickerView:(UIPickerView *)arg1 widthForComponent:(int)arg2;
+- (void)pickerView:(UIPickerView *)arg1 didSelectRow:(long long)arg2 inComponent:(long long)arg3;
+- (UIView *)pickerView:(UIPickerView *)arg1 viewForRow:(long long)arg2 forComponent:(long long)arg3 reusingView:(UIView *)arg4;
+- (NSAttributedString *)pickerView:(UIPickerView *)arg1 attributedTitleForRow:(long long)arg2 forComponent:(long long)arg3;
+- (NSString *)pickerView:(UIPickerView *)arg1 titleForRow:(long long)arg2 forComponent:(long long)arg3;
+- (double)pickerView:(UIPickerView *)arg1 rowHeightForComponent:(long long)arg2;
+- (double)pickerView:(UIPickerView *)arg1 widthForComponent:(long long)arg2;
 @end
