+++ UIKit.framework/iPhoneOS.platform/UITableViewVisibleCells.h
@@ -12,13 +12,13 @@ __attribute__((visibility("hidden")))
     NSMutableArray *_array;
 }
 
-- (void)replaceObjectAtIndex:(unsigned int)arg1 withObject:(id)arg2;
+- (void)replaceObjectAtIndex:(unsigned long long)arg1 withObject:(id)arg2;
 - (void)removeLastObject;
 - (void)addObject:(id)arg1;
-- (void)removeObjectAtIndex:(unsigned int)arg1;
-- (void)insertObject:(id)arg1 atIndex:(unsigned int)arg2;
-- (id)objectAtIndex:(unsigned int)arg1;
-- (unsigned int)count;
+- (void)removeObjectAtIndex:(unsigned long long)arg1;
+- (void)insertObject:(id)arg1 atIndex:(unsigned long long)arg2;
+- (id)objectAtIndex:(unsigned long long)arg1;
+- (unsigned long long)count;
 - (void)dealloc;
 - (id)_array;
