+++ UIKit.framework/iPhoneOS.platform/_UIDatePickerMode_MonthDay.h
@@ -11,10 +11,10 @@ __attribute__((visibility("hidden")))
 {
 }
 
-+ (unsigned int)extractableCalendarUnits;
-+ (int)datePickerMode;
-- (unsigned int)nextUnitSmallerThanUnit:(unsigned int)arg1;
-- (unsigned int)nextUnitLargerThanUnit:(unsigned int)arg1;
++ (unsigned long long)extractableCalendarUnits;
++ (long long)datePickerMode;
+- (unsigned long long)nextUnitSmallerThanUnit:(unsigned long long)arg1;
+- (unsigned long long)nextUnitLargerThanUnit:(unsigned long long)arg1;
 
 @end
