+++ UIKit.framework/iPhoneOS.platform/_UISlidingPopoverLayoutInfo.h
@@ -12,7 +12,7 @@ __attribute__((visibility("hidden")))
 }
 
 - (void)_updateOutputs;
-- (struct CGSize)_popoverViewSizeForContentSize:(struct CGSize)arg1 arrowDirection:(unsigned int)arg2;
+- (struct CGSize)_popoverViewSizeForContentSize:(struct CGSize)arg1 arrowDirection:(unsigned long long)arg2;
 - (id)init;
 
 @end
