+++ UIKit.framework/iPhoneSimulator.platform/UIActionSheetDelegate-Protocol.h
@@ -11,11 +11,11 @@
 @protocol UIActionSheetDelegate <NSObject>
 
 @optional
-- (void)actionSheet:(UIActionSheet *)arg1 didDismissWithButtonIndex:(int)arg2;
-- (void)actionSheet:(UIActionSheet *)arg1 willDismissWithButtonIndex:(int)arg2;
+- (void)actionSheet:(UIActionSheet *)arg1 didDismissWithButtonIndex:(long long)arg2;
+- (void)actionSheet:(UIActionSheet *)arg1 willDismissWithButtonIndex:(long long)arg2;
 - (void)didPresentActionSheet:(UIActionSheet *)arg1;
 - (void)willPresentActionSheet:(UIActionSheet *)arg1;
 - (void)actionSheetCancel:(UIActionSheet *)arg1;
-- (void)actionSheet:(UIActionSheet *)arg1 clickedButtonAtIndex:(int)arg2;
+- (void)actionSheet:(UIActionSheet *)arg1 clickedButtonAtIndex:(long long)arg2;
 @end
