+++ UIKit.framework/iPhoneSimulator.platform/UICompositeImageView.h
@@ -16,9 +16,9 @@ __attribute__((visibility("hidden")))
 
 - (void)drawRect:(struct CGRect)arg1;
 - (void)removeAllImages;
-- (void)addImage:(id)arg1 toRect:(struct CGRect)arg2 fromRect:(struct CGRect)arg3 operation:(int)arg4 fraction:(float)arg5;
+- (void)addImage:(id)arg1 toRect:(struct CGRect)arg2 fromRect:(struct CGRect)arg3 operation:(int)arg4 fraction:(double)arg5;
 - (void)addImage:(id)arg1 toRect:(struct CGRect)arg2 fromRect:(struct CGRect)arg3;
-- (void)addImage:(id)arg1 operation:(int)arg2 fraction:(float)arg3;
+- (void)addImage:(id)arg1 operation:(int)arg2 fraction:(double)arg3;
 - (void)addImage:(id)arg1;
 - (void)dealloc;
 - (id)initWithFrame:(struct CGRect)arg1;
