+++ UIKit.framework/iPhoneSimulator.platform/UIDocumentAlertPresenter.h
@@ -20,7 +20,7 @@ __attribute__((visibility("hidden")))
 
 + (id)_presentAlertWithError:(id)arg1 completionHandler:(CDUnknownBlockType)arg2;
 - (void)alertViewCancel:(id)arg1;
-- (void)alertView:(id)arg1 clickedButtonAtIndex:(int)arg2;
+- (void)alertView:(id)arg1 clickedButtonAtIndex:(long long)arg2;
 - (void)_forceFinishNow;
 - (void)dealloc;
 - (id)initWithError:(id)arg1 completionHandler:(CDUnknownBlockType)arg2;
