+++ UIKit.framework/iPhoneSimulator.platform/_UIPageViewControllerContentView.h
@@ -21,7 +21,7 @@ __attribute__((visibility("hidden")))
 - (void)layoutSubviews;
 - (void)_setupPageControl:(id)arg1;
 - (struct CGRect)_scrollViewFrame;
-- (float)_pageSpacing;
+- (double)_pageSpacing;
 - (void)setFrame:(struct CGRect)arg1;
 - (void)setBounds:(struct CGRect)arg1;
 - (void)invalidatePageViewController;
