+++ UIKit.framework/iPhoneSimulator.platform/_UITouchForwardingRecipient.h
@@ -13,12 +13,12 @@ __attribute__((visibility("hidden")))
 {
     UIResponder *fromResponder;
     UIResponder *responder;
-    int recordedPhase;
-    int autocompletedPhase;
+    long long recordedPhase;
+    long long autocompletedPhase;
 }
 
-@property(nonatomic) int autocompletedPhase; // @synthesize autocompletedPhase;
-@property(nonatomic) int recordedPhase; // @synthesize recordedPhase;
+@property(nonatomic) long long autocompletedPhase; // @synthesize autocompletedPhase;
+@property(nonatomic) long long recordedPhase; // @synthesize recordedPhase;
 @property(nonatomic) UIResponder *responder; // @synthesize responder;
 @property(nonatomic) UIResponder *fromResponder; // @synthesize fromResponder;
 - (id)description;
