+++ UIKit.framework/iPhoneSimulator.platform/_UIViewServiceDeputyRotationDelegate-Protocol.h
@@ -10,8 +10,8 @@
 
 @protocol _UIViewServiceDeputyRotationDelegate <NSObject>
 - (void)finishRotation;
-- (void)rotateToInterfaceOrientation:(int)arg1;
-- (void)willRotateToInterfaceOrientation:(int)arg1;
+- (void)rotateToInterfaceOrientation:(long long)arg1;
+- (void)willRotateToInterfaceOrientation:(long long)arg1;
 - (void)setHostedWindow:(UIWindow *)arg1;
 @end
