+++ UIKit.framework/iPhoneSimulator.platform/_UIViewServiceFencingController.h
@@ -14,14 +14,14 @@ __attribute__((visibility("hidden")))
     int _lock;
     NSMutableSet *_pendingFenceSendRights;
     NSMutableArray *_resumeActions;
-    unsigned int _expectedParticipatingFencingProxyCount;
+    unsigned long long _expectedParticipatingFencingProxyCount;
     NSTimer *_fencingControlTimeoutTimer;
 }
 
 + (id)activeFencePort;
 - (void)_fencingControlTimedOut;
 - (void)fencingControlProxy:(id)arg1 didEndFencingWithSendRight:(id)arg2;
-- (void)fencingControlProxy:(id)arg1 didBeginFencingWithSendRight:(id)arg2 expectedParticipatingFencingProxyCount:(unsigned int)arg3;
+- (void)fencingControlProxy:(id)arg1 didBeginFencingWithSendRight:(id)arg2 expectedParticipatingFencingProxyCount:(unsigned long long)arg3;
 - (void)dealloc;
 - (id)init;
